#1 - Unmodified import of Course Builder 1.6.0
I went to
https://code.google.com/p/course-builder/wiki/DownloadCourseBuilder

Downloaded their .zip file and put it into our repository.

Note, this is not a git fork but a download as that was the recommended
approach.  Not being a fork may lead to challenges in the future as
Google updates to 1.6.1 and beyond.  However, a fork/clone does not
work easily due to missing 3rd party code.  The download version brings
in all the libraries needed, a clone alone would not compile as all of
the 3rd party code would be missing.  There are about 11 3rd party
packages and a download is the easiest way to get them all.

Note, current code is untested.  Content will be about a Google Search
course.